start refactoring dev command
